requirements.txt
Pillow==8.0.1
pyaes==1.6.1
pyasn1==0.4.8
rsa==4.6
Telethon==1.17.5
